+++ bluetooth_low_energy_android/example/lib/types.dart
@@ -1,0 +1,10 @@
+export 'package:bluetooth_low_energy_platform_interface/bluetooth_low_energy_platform_interface.dart'
+    hide
+        CentralManager,
+        PeripheralManager,
+        MyObject,
+        MyCentral,
+        MyPeripheral,
+        MyGattService,
+        MyGattCharacteristic,
+        MyGattDescriptor;